requirements.txt
fastapi
uvicorn
pyjwt
python-jose[cryptography] 
passlib[bcrypt]
ultralytics
